chore: prepare tasks representing input handler
